refactor(config): move cache related config into own file
